Update cube2.py
Initially I didn't care for the code to be fast because the animation looked worse if the frames were generated faster. It was slow enough for me to not have to use time and add a delay. But after I added another direction of rotation, the animation became too slow. So i simply precomputed the sines and cosines and their squares instead of calculating it on the spot every iteration to make it faster.